requirements.txt
python-telegram-bot==13.15
fpdf==1.7.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
